class for sparql insert
